# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "API REST de Lebane para gestionar departamentos en venta: alta, edición, listado con filtros, fotos y consultas de interesados.\n\n**Convenciones**\n\n- **Errores**: toda respuesta 4xx/5xx tiene el mismo esquema `ApiError`, con un código estable en `error`, un mensaje apto para mostrar al usuario y el `requestId` para soporte. Nunca incluye stack traces, SQL ni detalles de infraestructura. Los errores de validación detallan cada campo en `fieldErrors` (las claves son las rutas del body o los nombres de los parámetros, p. ej. `direccion.ciudad`).\n- **Correlation ID**: cada request puede enviar `X-Request-Id`; si no lo envía o no es válido, el backend genera uno. La respuesta siempre lo devuelve y queda en todos los logs del request.\n- **Concurrencia optimista**: el detalle, el alta y la edición devuelven `ETag` con la versión del departamento. Enviarlo en `If-Match` al editar evita pisar cambios de otro usuario (412 si cambió).\n- **Paginación**: `page` (desde 0) y `size` (1 a 100), con una ventana máxima de 10.000 resultados por búsqueda. La respuesta usa el formato estándar de Spring Data (`content` + `page`).\n- **Disponibilidad**: si una dependencia no crítica falla (storage de imágenes, proveedor de direcciones), solo se degradan las operaciones que la usan: 503 `STORAGE_UNAVAILABLE` al subir fotos, respuesta `degradado: true` en el autocompletado. El resto de la API sigue funcionando.\n",
...
    "version" : "v1"
...
  "servers" : [ {
...
  "paths" : {
...
    "/api/v1/departamentos/{departamentoId}/imagenes/{imagenId}" : {
      "delete" : {
        "description" : "Elimina la foto del departamento. Si el storage no está disponible, la foto se elimina igual (deja de verse en la aplicación) y el archivo queda registrado en los logs como huérfano para limpiarlo después. Un departamento VENDIDO no admite cambios en sus fotos (409 `DEPARTAMENTO_NO_DISPONIBLE`).",
        "operationId" : "eliminar",
        "parameters" : [ {
          "description" : "ID del departamento",
          "example" : 1,
          "in" : "path",
          "name" : "departamentoId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "exclusiveMinimum" : 0
          }
        }, {
          "description" : "ID de la foto",
          "example" : 19,
          "in" : "path",
          "name" : "imagenId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "exclusiveMinimum" : 0
          }
        }, {
          "description" : "Correlation ID del request (1 a 128 caracteres: letras, números, '.', '_', ':' o '-'). Si falta o no es válido, el backend genera un UUID.",
          "in" : "header",
          "name" : "X-Request-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "7f3c2a9e-req-42",
            "pattern" : "^[A-Za-z0-9._:\\-]{1,128}$"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Foto eliminada",
            "headers" : {
              "X-Request-Id" : {
                "description" : "Correlation ID del request (el recibido o uno generado)",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalError"
          },
          "503" : {
            "$ref" : "#/components/responses/ServiceUnavailable"
          }
        },
        "summary" : "Eliminar una foto",
        "tags" : [ "Imágenes" ]
      }
    },
...
  },
  "components" : {
    "responses" : {
      "BadRequest" : {
        "content" : {
          "application/json" : {
            "examples" : {
              "json" : {
                "summary" : "JSON mal formado",
                "value" : {
                  "error" : "BAD_REQUEST",
                  "message" : "El cuerpo de la solicitud no es JSON válido",
                  "path" : "/api/v1/departamentos",
                  "requestId" : "7f3c2a9e-req-42",
                  "status" : 400,
                  "timestamp" : "2026-10-01T12:00:00Z"
                }
              },
              "validacion" : {
                "summary" : "Validación de campos",
                "value" : {
                  "error" : "VALIDATION_ERROR",
                  "fieldErrors" : {
                    "dormitorios" : "debe ser menor que la cantidad de ambientes",
                    "titulo" : "no debe estar vacío"
                  },
                  "message" : "La solicitud contiene datos inválidos",
                  "path" : "/api/v1/departamentos",
                  "requestId" : "7f3c2a9e-req-42",
                  "status" : 400,
                  "timestamp" : "2026-10-01T12:00:00Z"
                }
              }
            },
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "Datos inválidos: body, parámetros o archivo. Cada campo con error se detalla en `fieldErrors`.",
        "headers" : {
          "X-Request-Id" : {
            "description" : "Correlation ID del request (el recibido o uno generado)",
            "schema" : {
              "type" : "string"
            }
          }
        }
      },
      "Conflict" : {
        "content" : {
          "application/json" : {
            "examples" : {
              "concurrencia" : {
                "summary" : "Modificación simultánea",
                "value" : {
                  "error" : "CONCURRENT_MODIFICATION",
                  "message" : "El recurso fue modificado por otra operación; recargalo y volvé a intentar",
                  "path" : "/api/v1/departamentos/7",
                  "requestId" : "7f3c2a9e-req-42",
                  "status" : 409,
                  "timestamp" : "2026-10-01T12:00:00Z"
                }
              },
              "limite" : {
                "summary" : "Límite de fotos alcanzado",
                "value" : {
                  "error" : "LIMITE_IMAGENES_ALCANZADO",
                  "message" : "El departamento ya tiene el máximo de 5 fotos",
                  "path" : "/api/v1/departamentos/7/imagenes",
                  "requestId" : "7f3c2a9e-req-42",
                  "status" : 409,
                  "timestamp" : "2026-10-01T12:00:00Z"
                }
              },
              "transicion" : {
                "summary" : "Cambio de estado no permitido",
                "value" : {
                  "error" : "TRANSICION_DE_ESTADO_INVALIDA",
                  "message" : "Un departamento no se puede publicar directamente como vendido",
                  "path" : "/api/v1/departamentos",
                  "requestId" : "7f3c2a9e-req-42",
                  "status" : 409,
                  "timestamp" : "2026-10-01T12:00:00Z"
                }
              },
              "vendido" : {
                "summary" : "Departamento vendido",
                "value" : {
                  "error" : "DEPARTAMENTO_NO_DISPONIBLE",
                  "message" : "El departamento ya no está disponible y no recibe nuevas consultas",
                  "path" : "/api/v1/departamentos/7/consultas",
                  "requestId" : "7f3c2a9e-req-42",
                  "status" : 409,
                  "timestamp" : "2026-10-01T12:00:00Z"
                }
              },
              "vendidoNoModificable" : {
                "summary" : "Departamento vendido: no se modifica",
                "value" : {
                  "error" : "DEPARTAMENTO_NO_DISPONIBLE",
                  "message" : "El departamento ya fue vendido y no se puede modificar",
                  "path" : "/api/v1/departamentos/7",
                  "requestId" : "7f3c2a9e-req-42",
                  "status" : 409,
                  "timestamp" : "2026-10-01T12:00:00Z"
                }
              }
            },
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "La operación entra en conflicto con el estado actual (ver `error`).",
        "headers" : {
          "X-Request-Id" : {
            "description" : "Correlation ID del request (el recibido o uno generado)",
            "schema" : {
              "type" : "string"
            }
          }
        }
      },
      "InternalError" : {
        "content" : {
          "application/json" : {
            "examples" : {
              "inesperado" : {
                "summary" : "Error inesperado",
                "value" : {
                  "error" : "INTERNAL_ERROR",
                  "message" : "Ocurrió un error inesperado; si persiste, informá el requestId",
                  "path" : "/api/v1/departamentos",
                  "requestId" : "7f3c2a9e-req-42",
                  "status" : 500,
                  "timestamp" : "2026-10-01T12:00:00Z"
                }
              }
            },
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "Error inesperado. El mensaje es genérico; el `requestId` permite encontrar el detalle en los logs.",
        "headers" : {
          "X-Request-Id" : {
            "description" : "Correlation ID del request (el recibido o uno generado)",
            "schema" : {
              "type" : "string"
            }
          }
        }
      },
      "NotFound" : {
        "content" : {
          "application/json" : {
            "examples" : {
              "departamento" : {
                "summary" : "Departamento inexistente",
                "value" : {
                  "error" : "NOT_FOUND",
                  "message" : "No se encontró el departamento solicitado",
                  "path" : "/api/v1/departamentos/999",
                  "requestId" : "7f3c2a9e-req-42",
                  "status" : 404,
                  "timestamp" : "2026-10-01T12:00:00Z"
                }
              }
            },
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "El recurso no existe.",
        "headers" : {
          "X-Request-Id" : {
            "description" : "Correlation ID del request (el recibido o uno generado)",
            "schema" : {
              "type" : "string"
            }
          }
        }
      },
...
      "ServiceUnavailable" : {
        "content" : {
          "application/json" : {
            "examples" : {
              "base" : {
                "summary" : "Base de datos no disponible",
                "value" : {
                  "error" : "SERVICE_UNAVAILABLE",
                  "message" : "El servicio no está disponible temporalmente; intentá nuevamente",
                  "path" : "/api/v1/departamentos",
                  "requestId" : "7f3c2a9e-req-42",
                  "status" : 503,
                  "timestamp" : "2026-10-01T12:00:00Z"
                }
              },
              "storage" : {
                "summary" : "Storage de imágenes no disponible",
                "value" : {
                  "error" : "STORAGE_UNAVAILABLE",
                  "message" : "El servicio de imágenes no está disponible; intentá nuevamente en unos minutos",
                  "path" : "/api/v1/departamentos/7/imagenes",
                  "requestId" : "7f3c2a9e-req-42",
                  "status" : 503,
                  "timestamp" : "2026-10-01T12:00:00Z"
                }
              }
            },
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            }
          }
        },
        "description" : "Una dependencia necesaria no está disponible (base de datos o, al subir fotos, el storage). Reintentar más tarde.",
        "headers" : {
          "X-Request-Id" : {
            "description" : "Correlation ID del request (el recibido o uno generado)",
            "schema" : {
              "type" : "string"
            }
          }
        }
      },
...
    },
    "schemas" : {
      "ApiError" : {
        "description" : "Error de la API. Mismo esquema para toda respuesta 4xx/5xx; nunca contiene stack traces, SQL ni detalles de infraestructura.",
        "properties" : {
          "error" : {
            "type" : "string",
            "description" : "Código estable del error. Los clientes pueden depender de él; el mensaje es solo informativo.",
            "enum" : [ "VALIDATION_ERROR", "BAD_REQUEST", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "METHOD_NOT_ALLOWED", "NOT_ACCEPTABLE", "CONFLICT", "CONCURRENT_MODIFICATION", "PRECONDITION_FAILED", "PAYLOAD_TOO_LARGE", "UNSUPPORTED_MEDIA_TYPE", "DEPARTAMENTO_NO_DISPONIBLE", "TRANSICION_DE_ESTADO_INVALIDA", "LIMITE_IMAGENES_ALCANZADO", "STORAGE_UNAVAILABLE", "SERVICE_UNAVAILABLE", "INTERNAL_ERROR" ]
          },
          "fieldErrors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string",
              "description" : "Errores por campo (solo en errores de validación). La clave es la ruta del campo en el body o el nombre del parámetro; se omite si no hay.",
              "example" : "{\"direccion.ciudad\":\"no debe estar vacío\"}"
            },
            "description" : "Errores por campo (solo en errores de validación). La clave es la ruta del campo en el body o el nombre del parámetro; se omite si no hay.",
            "example" : {
              "direccion.ciudad" : "no debe estar vacío"
            }
          },
          "message" : {
            "type" : "string",
            "description" : "Mensaje apto para mostrar al usuario",
            "example" : "La solicitud contiene datos inválidos"
          },
          "path" : {
            "type" : "string",
            "description" : "Path del request (sin query string)",
            "example" : "/api/v1/departamentos"
          },
          "requestId" : {
            "type" : "string",
            "description" : "Correlation ID del request: permite encontrar el detalle en los logs",
            "example" : "7f3c2a9e-req-42"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Status HTTP",
            "example" : 400
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "Momento del error (UTC)",
            "example" : "2026-10-01T12:00:00Z"
          }
        },
        "required" : [ "error", "status", "timestamp" ]
      },
...
    }
  }
}